+++ src/GameState.hpp
@@ -11,7 +11,7 @@ public:
     GameState();
 
     void onInit() override;
-    void onUpdate(double dt) override;
+    void onUpdate(float dt) override;
     void onDraw(sf::RenderTarget& target) override;
     void onHandleEvent(sf::Event& event) override;
